MIP-262 refactored client package under api to remove warning on deployment
